+++ test/contrib/hadoop_test.py
